// src/models/public/CheckValues.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { AccountsAccountId } from './Accounts';
import type { ChecksCheckId } from './Checks';
import type { UnitsUnitId } from './Units';

/** Identifier type for public.check_values */
export type CheckValuesCheckValueId = string & { __brand: 'public.check_values' };

/**
 * Represents the table public.check_values
 * value-ing checks i.e. the situation of the subproject in this place
 */
export default interface CheckValues {
  check_value_id: CheckValuesCheckValueId;

  /** redundant account_id enhances data safety */
  account_id: AccountsAccountId | null;

  check_id: ChecksCheckId | null;

  unit_id: UnitsUnitId | null;

  /** Used for integer values */
  value_integer: number | null;

  /** Used for numeric values */
  value_numeric: number | null;

  /** Used for text values */
  value_text: string | null;

  label: string | null;

  created_at: Date;

  updated_at: Date;

  updated_by: string | null;
}

/**
 * Represents the initializer for the table public.check_values
 * value-ing checks i.e. the situation of the subproject in this place
 */
export interface CheckValuesInitializer {
  /** Default value: uuid_generate_v7() */
  check_value_id?: CheckValuesCheckValueId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  check_id?: ChecksCheckId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer values */
  value_integer?: number | null;

  /** Used for numeric values */
  value_numeric?: number | null;

  /** Used for text values */
  value_text?: string | null;

  label?: string | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;

  updated_by?: string | null;
}

/**
 * Represents the mutator for the table public.check_values
 * value-ing checks i.e. the situation of the subproject in this place
 */
export interface CheckValuesMutator {
  check_value_id?: CheckValuesCheckValueId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  check_id?: ChecksCheckId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer values */
  value_integer?: number | null;

  /** Used for numeric values */
  value_numeric?: number | null;

  /** Used for text values */
  value_text?: string | null;

  label?: string | null;

  created_at?: Date;

  updated_at?: Date;

  updated_by?: string | null;
}